+++ python_modules/dagster/dagster_tests/execution_tests/test_logical_versions.py
@@ -16,9 +16,7 @@ from dagster import (
 from dagster._config.field import Field
 from dagster._core.definitions.asset_graph import AssetGraph
 from dagster._core.definitions.asset_out import AssetOut
-from dagster._core.definitions.decorators.asset_decorator import multi_asset
-from dagster._core.definitions.events import AssetKey, Output
-from dagster._core.definitions.logical_version import (
+from dagster._core.definitions.data_version import (
     CODE_VERSION_TAG_KEY,
     INPUT_LOGICAL_VERSION_TAG_KEY_PREFIX,
     LOGICAL_VERSION_TAG_KEY,
@@ -28,6 +26,8 @@ from dagster._core.definitions.logical_version import (
     StaleStatus,
     compute_logical_version,
 )
+from dagster._core.definitions.decorators.asset_decorator import multi_asset
+from dagster._core.definitions.events import AssetKey, Output
 from dagster._core.definitions.observe import observe
 from dagster._core.definitions.partition import StaticPartitionsDefinition
 from dagster._core.execution.context.compute import OpExecutionContext
